+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -2,12 +2,17 @@ package roomescape.reservation.controller;
 
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.reservation.dto.ReservationCreateRequest;
 import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.dto.WaitingResponse;
 import roomescape.reservation.service.ReservationService;
-
 import java.time.LocalDate;
 import java.util.List;
 
@@ -26,40 +31,22 @@ public class AdminReservationController {
         return reservationService.createReservation(request);
     }
 
-    @PostMapping("/waitings/{id}")
-    public ReservationResponse approveWaiting(@PathVariable Long id) {
-        return reservationService.approveWaiting(id);
-    }
-
     @GetMapping("/reservations")
     public List<ReservationResponse> readReservations() {
         return reservationService.readReservations();
     }
 
     @GetMapping("/reservations/search")
-    public List<ReservationResponse> readReservations(
-            @RequestParam LocalDate dateFrom,
-            @RequestParam LocalDate dateTo,
-            @RequestParam Long memberId,
-            @RequestParam Long themeId
-    ) {
+    public List<ReservationResponse> readReservations(@RequestParam LocalDate dateFrom,
+                                                      @RequestParam LocalDate dateTo,
+                                                      @RequestParam Long memberId,
+                                                      @RequestParam Long themeId) {
         return reservationService.searchReservations(dateFrom, dateTo, memberId, themeId);
-    }
-
-    @GetMapping("/waitings")
-    public List<WaitingResponse> readWaitings() {
-        return reservationService.readWaitings();
     }
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
         reservationService.deleteReservation(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @DeleteMapping("/waitings/{id}")
-    public ResponseEntity<Void> rejectWaiting(@PathVariable Long id) {
-        reservationService.deleteWaiting(id);
         return ResponseEntity.noContent().build();
     }
 }
