+++ src/main/java/roomescape/reservation/controller/MemberWaitingController.java
@@ -1,37 +1,38 @@
 package roomescape.reservation.controller;
 
 import jakarta.validation.Valid;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.dto.LoginMember;
 import roomescape.reservation.dto.MemberReservationCreateRequest;
-import roomescape.reservation.dto.MemberReservationResponse;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.WaitingResponse;
 import roomescape.reservation.service.ReservationService;
-import java.util.List;
 
 @RestController
-@RequestMapping("/reservations")
-public class MemberReservationController {
+@RequestMapping("/waitings")
+public class MemberWaitingController {
 
     private final ReservationService reservationService;
 
-    public MemberReservationController(ReservationService reservationService) {
+    public MemberWaitingController(ReservationService reservationService) {
         this.reservationService = reservationService;
     }
 
     @PostMapping
-    public ReservationResponse createReservation(
+    public WaitingResponse createWaitingReservation(
             @Valid @RequestBody MemberReservationCreateRequest request,
             LoginMember member) {
-        return reservationService.createReservation(request, member);
+        return reservationService.createWaitingReservation(request, member);
     }
 
-    @GetMapping("/mine")
-    public List<MemberReservationResponse> readMemberReservations(LoginMember loginMember) {
-        return reservationService.readMemberReservations(loginMember);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteWaitingReservation(@PathVariable Long id, LoginMember loginMember) {
+        reservationService.deleteWaitingReservation(id, loginMember);
+        return ResponseEntity.noContent().build();
     }
 }
